feat: add radar bubble tooltips with keyboard accessibility
Replaces native SVG <title> browser tooltips on unlabeled radar bubbles
with a styled, mobile-friendly HTML overlay. Extracts RadarChart into
RadarChartClient (client component) to support hover/focus state.

- Every bubble: role="button", tabIndex=0, aria-label="Name — N roles, Category"
- Hover and keyboard focus both show the tooltip (name + count + category)
- Enter / Space navigate to /?tech=<name>
- Visible accent focus ring (SVG circle) on keyboard focus
- Expanded transparent hit target for small bubbles / touch
- Inline styles for tooltip to avoid Turbopack CSS-class loading issues in dev
- Category leader inline labels preserved unchanged

Closes #29
