+++ frontend/app/globals.css
@@ -210,3 +210,4 @@ body {
     width: 30% !important; /* static partial fill — visible but not moving */
   }
 }
+
